Initial Commit
I am not a fan of LDAP.  I’m sure there are CLI tools that can do this
same thing, but they’re not cross platform, and I can never seem to get
the working correctly.

Putting this out in the ether to have a baseline to work from the next
time I need to do something like this.